+++ packages/components/calendar/hooks/useColHeaders.ts
@@ -1,9 +1,9 @@
 import { computed } from 'vue';
 
-import { CellColHeader, CalendarState } from '../interface';
+import { CellColHeader, CalendarState } from '../types';
 import { TdCalendarProps, WeekDay } from '../type';
 import { useConfig } from '../../hooks/useConfig';
-import { COMPONENT_NAME } from '../const';
+import { COMPONENT_NAME } from '../consts';
 import * as utils from '../utils';
 import { isObject } from 'lodash-es';
